# backend/src/main/resources/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: Anki
  description: Anki 작업
  version: 1.0.0
servers:
- url: http://localhost:8080/
  description: 8080 포트에서 사용
- url: http://localhost:7070/
  description: 7070 포트에서 사용
- url: https://virtserver.swaggerhub.com/js990311/Anki/1.0.0
  description: swagger-hub에서 사용
...
paths:
...
  /decks:
    get:
      tags:
      - Decks
      summary: 검색어 조건에 맞는 Deck 보기
      description: 검색어 조건에 맞는 Deck 보기
      operationId: deck-controller-test/get-decks/
      parameters:
      - name: queryType
        in: query
        description: "난이도 기준 검색인지, 의미상 구분에 따른 검색인지"
        required: true
        schema:
          type: string
          enum:
          - level
          - meaning
      responses:
        "200":
          description: "200"
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: "#/components/schemas/decks"
              examples:
                deck-controller-test/get-decks/level:
                  value: "{\"size\":3,\"content\":[{\"category\":\"easy\",\"cardCounts\"\
                    :20,\"overdueCounts\":20,\"maturitiyCounts\":0,\"overdueRate\"\
                    :1.0,\"maturitiyRate\":0.0},{\"category\":\"hard\",\"cardCounts\"\
                    :20,\"overdueCounts\":20,\"maturitiyCounts\":0,\"overdueRate\"\
                    :1.0,\"maturitiyRate\":0.0},{\"category\":\"normal\",\"cardCounts\"\
                    :20,\"overdueCounts\":20,\"maturitiyCounts\":0,\"overdueRate\"\
                    :1.0,\"maturitiyRate\":0.0}]}"
                deck-controller-test/get-decks/meaning:
                  value: "{\"size\":27,\"content\":[{\"category\":\"ACTION\",\"cardCounts\"\
                    :2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"overdueRate\":1.0,\"\
                    maturitiyRate\":0.0},{\"category\":\"ADMINISTRATION\",\"cardCounts\"\
                    :2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"overdueRate\":1.0,\"\
                    maturitiyRate\":0.0},{\"category\":\"COMMUNICATION\",\"cardCounts\"\
                    :2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"overdueRate\":1.0,\"\
                    maturitiyRate\":0.0},{\"category\":\"CONCEPT\",\"cardCounts\"\
                    :2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"overdueRate\":1.0,\"\
                    maturitiyRate\":0.0},{\"category\":\"CULTURE\",\"cardCounts\"\
                    :2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"overdueRate\":1.0,\"\
                    maturitiyRate\":0.0},{\"category\":\"ECONOMY\",\"cardCounts\"\
                    :3,\"overdueCounts\":3,\"maturitiyCounts\":0,\"overdueRate\":1.0,\"\
                    maturitiyRate\":0.0},{\"category\":\"FASHION_AND_APPEARANCE\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"FEELING\"\
                    ,\"cardCounts\":3,\"overdueCounts\":3,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"FOOD\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"GRAMMAR_AND_LANGUAGE\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"HOME\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"HOSPITAL\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"LIFE\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"LIVING\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"NATURE\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"NEWS\"\
                    ,\"cardCounts\":3,\"overdueCounts\":3,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"NUMBER\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"PERSONALITY\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"POLITICS\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"RELATIONSHIPS\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"RELIGION\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"SCHOOL\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"SCIENCE\"\
                    ,\"cardCounts\":3,\"overdueCounts\":3,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"TIME\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"TRANSPORT\"\
                    ,\"cardCounts\":3,\"overdueCounts\":3,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"WEATHER\"\
                    ,\"cardCounts\":3,\"overdueCounts\":3,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0},{\"category\":\"WORK\"\
                    ,\"cardCounts\":2,\"overdueCounts\":2,\"maturitiyCounts\":0,\"\
                    overdueRate\":1.0,\"maturitiyRate\":0.0}]}"
...
components:
  schemas:
...
    decks:
      title: decks
      type: object
      properties:
        size:
          type: number
          description: content의 크기
        content:
          type: array
          description: 실제 데이터들
          items:
            type: object
            properties:
              overdueRate:
                type: number
                description: due가 지난 상태인 카드 비율
              cardCounts:
                type: number
                description: 덱에 포함된 카드 개수
              overdueCounts:
                type: number
                description: due가 지난 상태인 카드 개수
              maturitiyRate:
                type: number
                description: "state가 review인 카드 비율:"
              category:
                type: string
                description: 카드 분류
              maturitiyCounts:
                type: number
                description: state가 review인 카드 개수
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
security:
  - cookieAuth: []